move shared stuff to utils
