+++ tests/deploy/api/test_batch_endpoints.py
@@ -4,7 +4,7 @@ from fastapi.testclient import TestClient
 from ragna.deploy import Config
 from ragna.deploy._api import app
 
-from .utils import authenticate
+from ..utils import authenticate
 
 
 def test_batch_sequential_upload_equivalence(tmp_local_root):
