--- tests/deploy/api/utils.py
@@ -1,23 +1,0 @@
-import os
-
-from fastapi.testclient import TestClient
-
-from ragna.core._utils import default_user
-
-
-def authenticate(client: TestClient) -> None:
-    username = default_user()
-    token = (
-        client.post(
-            "/token",
-            data={
-                "username": username,
-                "password": os.environ.get(
-                    "RAGNA_DEMO_AUTHENTICATION_PASSWORD", username
-                ),
-            },
-        )
-        .raise_for_status()
-        .json()
-    )
-    client.headers["Authorization"] = f"Bearer {token}"
